refactor: create UserDao interface
