+++ slipp/src/main/java/slipp/dao/UserDao.java
@@ -1,53 +1,15 @@
 package slipp.dao;
 
-import nextstep.jdbc.JdbcTemplate;
-import nextstep.jdbc.RowMapper;
-import nextstep.stereotype.Repository;
 import slipp.domain.User;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.List;
 
-@Repository
-public class UserDao {
-    private JdbcTemplate jdbcTemplate = JdbcTemplate.getInstance();
+public interface UserDao {
+    void insert(User user);
 
-    public void insert(User user) {
-        String sql = "INSERT INTO USERS VALUES (?, ?, ?, ?)";
-        jdbcTemplate.update(sql, user.getUserId(), user.getPassword(), user.getName(), user.getEmail());
-    }
+    User findByUserId(String userId);
 
-    public User findByUserId(String userId) {
-        String sql = "SELECT userId, password, name, email FROM USERS WHERE userid=?";
+    void update(User user);
 
-        RowMapper<User> rm = new RowMapper<User>() {
-            @Override
-            public User mapRow(ResultSet rs) throws SQLException {
-                return new User(rs.getString("userId"), rs.getString("password"), rs.getString("name"),
-                        rs.getString("email"));
-            }
-        };
-
-        return jdbcTemplate.queryForObject(sql, rm, userId);
-    }
-
-    public List<User> findAll() throws SQLException {
-        String sql = "SELECT userId, password, name, email FROM USERS";
-
-        RowMapper<User> rm = new RowMapper<User>() {
-            @Override
-            public User mapRow(ResultSet rs) throws SQLException {
-                return new User(rs.getString("userId"), rs.getString("password"), rs.getString("name"),
-                        rs.getString("email"));
-            }
-        };
-
-        return jdbcTemplate.query(sql, rm);
-    }
-
-    public void update(User user) {
-        String sql = "UPDATE USERS set password = ?, name = ?, email = ? WHERE userId = ?";
-        jdbcTemplate.update(sql, user.getPassword(), user.getName(), user.getEmail(), user.getUserId());
-    }
+    List<User> findAll();
 }
