+++ slipp/src/test/java/slipp/dao/JdbcUserDaoTest.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class UserDaoTest {
+public class JdbcUserDaoTest {
     @BeforeEach
     public void setup() {
         ResourceDatabasePopulator populator = new ResourceDatabasePopulator();
@@ -24,21 +24,21 @@ public class UserDaoTest {
     @Test
     public void crud() throws Exception {
         User expected = new User("userId", "password", "name", "javajigi@email.com");
-        UserDao userDao = new UserDao();
-        userDao.insert(expected);
-        User actual = userDao.findByUserId(expected.getUserId());
+        JdbcUserDao jdbcUserDao = new JdbcUserDao();
+        jdbcUserDao.insert(expected);
+        User actual = jdbcUserDao.findByUserId(expected.getUserId());
         assertThat(actual).isEqualTo(expected);
 
         expected.update(new UserUpdatedDto("password2", "name2", "sanjigi@email.com"));
-        userDao.update(expected);
-        actual = userDao.findByUserId(expected.getUserId());
+        jdbcUserDao.update(expected);
+        actual = jdbcUserDao.findByUserId(expected.getUserId());
         assertThat(actual).isEqualTo(expected);
     }
 
     @Test
     public void findAll() throws Exception {
-        UserDao userDao = new UserDao();
-        List<User> users = userDao.findAll();
+        JdbcUserDao jdbcUserDao = new JdbcUserDao();
+        List<User> users = jdbcUserDao.findAll();
         assertThat(users).hasSize(1);
     }
 }
